4.1.5. Рефакторинг сложных компонентов
- Формирование разметки дополнительных опций и списка типов поездки вынесены в отдельный скрипт
