+++ src/mock/event-types.js
@@ -9,3 +9,4 @@ export const eventTypes = () => ([
   'sightseeing',
   'restaurant'
 ]);
+
