+++ src/mock/trip-event.js
@@ -173,3 +173,4 @@ export const generateTripEvent = () => {
     isBeingEdited: false
   };
 };
+
