+++ src/utils/render.js
@@ -33,3 +33,4 @@ export const createElement = (template) => {
 
   return newElement.firstChild;
 };
+
