+++ src/view/event-edit-view.js
@@ -2,56 +2,16 @@ import dayjs from 'dayjs';
 import {locations} from '../mock/locations';
 import {eventTypes} from '../mock/event-types';
 import AbstractView from './abstract-view';
+import {createEventTypesMarkup, createOffersSectionMarkup} from '../utils/forms';
 
 const createEventEditTemplate = (tripEvent) => {
   const {eventType, price, location, startDate, endDate, offers, description} = tripEvent;
   const startDatetime = dayjs(startDate).format('DD/MM/YY HH:mm ');
   const endDatetime = dayjs(endDate).format('DD/MM/YY HH:mm');
-
-  const createOfferMarkup = (offer) => {
-    const isChecked = offer.isChosen ? ' checked=""' : '';
-    const offerName = offer.name;
-    const offerPrice = offer.price;
-    const offerType = offer.type;
-
-    return `<div class="event__available-offers">
-                      <div class="event__offer-selector">
-                        <input class="event__offer-checkbox  visually-hidden" id="event-offer-${offerType}-1" type="checkbox" name="event-offer-${offerType}"${isChecked}>
-                        <label class="event__offer-label" for="event-offer-name-1">
-                          <span class="event__offer-title">${offerName}</span>
-                          +€&nbsp;
-                          <span class="event__offer-price">${offerPrice}</span>
-                        </label>
-                      </div>
-    `;
-  };
-  const createOffersListMarkup = (editedOffers) => {
-    if (editedOffers.length !== 0){
-      return `<section class="event__section  event__section--offers">
-                    <h3 class="event__section-title  event__section-title--offers">Offers</h3>${editedOffers}</section>`;
-    }
-
-    return '';
-  };
-  const createLocationOption = (city) => (`<option value="${city}"></option>`);
-  const createEventTypesMarkup = (types = eventTypes(), chosenEventType) => {
-    const createType = (currentType) => {
-      const isChecked = currentType === chosenEventType ? 'checked=""' : '';
-      const label = currentType.charAt(0).toUpperCase() + currentType.slice(1);
-      return `<div class="event__type-item">
-                          <input id="event-type-${currentType}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${currentType}" ${isChecked}>
-                          <label class="event__type-label  event__type-label--${currentType}" for="event-type-${currentType}-1">${label}</label>
-                        </div>`;
-    };
-
-    return types.map(createType).join('');
-  };
-
-  const editedOffersMarkup = offers.map(createOfferMarkup).join('');
-  const offersListMarkup = createOffersListMarkup(editedOffersMarkup);
-  const locationOptions = locations().map(createLocationOption).join('');
-  const eventTypesMarkup = createEventTypesMarkup(eventTypes(), eventType);
+  const locationOptions = locations().map((x) => (`<option value="${x}"></option>`)).join('');
   const eventTypeLabel = eventType.charAt(0).toUpperCase() + eventType.slice(1);
+  const editedOffersMarkup = createOffersSectionMarkup(offers);
+  const eventTypesMarkup = createEventTypesMarkup(eventTypes(), eventType);
 
   return `<li class="trip-events__item">
               <form class="event event--edit" action="#" method="post">
@@ -103,7 +63,7 @@ const createEventEditTemplate = (tripEvent) => {
                     <span class="visually-hidden">Open event</span>
                   </button>
                 </header>
-                <section class="event__details">${offersListMarkup}<section class="event__section  event__section--destination">
+                <section class="event__details">${editedOffersMarkup}<section class="event__section  event__section--destination">
                     <h3 class="event__section-title  event__section-title--destination">Destination</h3>
                     <p class="event__destination-description">${description}</p>
                   </section>
@@ -144,3 +104,4 @@ export default class EventEditView extends AbstractView {
     this._callback.rollupClick();
   }
 }
+
